requirements.txt
django
djangorestframework
mongoengine
django-rest-framework-mongoengine